// src/app/modules/master-forms/Models/VendorModel.ts
export class VendorModel {
    VendorId: number = 0;
    BranchId: number = 0;
    VendorName: string = '';
    Email: string = '';
    ContactNumber: string = '';
    PANNumber: string = '';
    IsGST: string = '';
    GSTNumber: string = '';
    City: string = '';
    Address: string = '';
    IsActive: string = '';
    Addedby: number = 0;
    CityName: string = '';
    AddedOn: string = '';
    LastUpdatedOn: string = '';
    Action: string = '';
}