// src/app/modules/master-forms/Models/CompanyList.ts
export class CompanyList {
  CompanyId: number = 0
  CompanyCode: string = '';
  CompanyName: string = '';
  CompanyAddress: string = '';
  CompanyCity: string = '';
  City: string = '';
  Pin: string = '';
  ContactNo: string = '';
  CompanyEmail: string = '';
  CompanyPAN: string = '';
  GSTNo: string = '';
  CityName: string = '';
  IsActive: string = '';
  Addedby: string = '';
  AddedOn: string = '';
  LastUpdatedOn: string = '';
  Action: string = '';
  IsPicklistAvailable: boolean;
}